+++ src/Tools/Logger/Reader/LogsMonitor.php
@@ -16,6 +16,23 @@ use InvalidArgumentException;
 interface LogsMonitor
 {
     /**
+     * Allows to redefine levels (if different from default).
+     * The list of Levels must be ordered from the least to the most significant.
+     * Instead of call this method, new levels can be set by pass them as second argument to filterByLevels.
+     *
+     * @param array $levels
+     * @return $this fluent interface
+     */
+    public function setLevels(array $levels);
+    
+    /**
+     * List of levels this LogsMonitor knows and understands.
+     *
+     * @return array
+     */
+    public function getLevels();
+    
+    /**
      * One can set minimum level of logs that will be streamed to OutputWriter by this LogsMonitor.
      * This is optional and if method is not call, all available logs are supposed to be streamed.
      *