+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All important changes to `fiisoft-logger` will be documented in this file
 
+## 2.0.0
+
+New methods (setLevels and getLevels) was added to interface LogsMonitor.
+
+Command ShowQueueLogsCommand can now handle option --show-levels and display available levels of logs.
+
 ## 1.4.0
 
 Updated dependencies. Method ShowQueueLogsCommand::handleInput() returns status code now (instead of exit).